remove polling; wait for modify events to diag file
Instead of constantly polling in a loop for any changes to the
`.bacon-locations`, and constantly pushing updates to the editor via
LSP, this will watch for changes to the `.bacon-locations` and only then
reevaluate the file and push diagnostics to the editor. This uses the
`notify` crate for cross-platform compatibility (it's used by
rust-analyzer itself, for example).

To avoid pushing too many updates at once the events are debounced using
`notify-debouncer-full`. The `syncronize_all_open_files_wait_millis`
configuration that was previously used as a wait time between polls is
now used as a timeout, ensuring that bacon-ls receives updates at least
this often. This allows us to move users silently over to the improved
implementation without requiring any new configuration or
re-configuration.